Fixing getPageViewsBetween
Hi again and happy new year.
I'm working with this package and problem with it is that getPageViewsBetween not working. I looked at codes in trait and I see the problem is making query for dates.
In code that you wrote, if both dates are available you override second 'where' and nothing work. so i propose new code to fix it. I think it give the idea to how to fix it if it's not how you code.
thanks